requirements.txt
aiohttp         3.6.2
async-timeout   3.0.1
attrs           20.1.0
certifi         2020.6.20
chardet         3.0.4
click           7.1.2
Cython          0.29.21
discord.py      1.4.1
Flask           1.1.2
idna            2.10
itsdangerous    1.1.0
Jinja2          2.11.2
MarkupSafe      1.1.1
memory-profiler 0.57.0
multidict       4.7.6
numpy           1.19.1
pip             20.1.1
psutil          5.7.2
pymongo         3.11.0
PyMySQL         0.10.0
requests        2.24.0
setuptools      47.1.0
urllib3         1.25.10
Werkzeug        1.0.1
yarl            1.5.1